intergrate and fix components

+++ SupermarketManagement/src/main/java/com/ou/repositories/LimitSaleRepository.java
@@ -53,7 +53,7 @@ public class LimitSaleRepository {
                 limitSale.setSaleId(resultSet.getInt("lsal_id"));
                 limitSale.setLsalFromDate(resultSet.getDate("lsal_from_date"));
                 limitSale.setLsalToDate(resultSet.getDate("lsal_to_date"));
-                salePercent.setSperPercent(resultSet.getFloat("sper_percent"));
+                salePercent.setSperPercent(resultSet.getInt("sper_percent"));
                 limitSale.setSalePercent(salePercent);
                 limitSales.add(limitSale);
             }
