+++ SupermarketManagement/src/main/java/com/ou/repositories/SaleRepositoryForTest.java
@@ -25,7 +25,7 @@ public class SaleRepositoryForTest {
                 sale.setSaleId(resultSet.getInt("sale_id"));
                 SalePercent salePercent = new SalePercent();
                 salePercent.setSperId(resultSet.getInt("sper_id"));
-                salePercent.setSperPercent(resultSet.getFloat("sper_percent"));
+                salePercent.setSperPercent(resultSet.getInt("sper_percent"));
                 salePercent.setSperIsActive(resultSet.getBoolean("sper_is_active"));
                 sale.setSalePercent(salePercent);
                 return sale;
